Update channel manager components with playlist and download support
- Implemented `download_manager` for handling artwork downloads.
- Added `channel_settings` for managing per-channel configurations.
- Updated `play_navigator` and `playlist_manager` to support new features.
- Modified `makapix_channel_impl` and `sdcard_channel_impl` to integrate new components.
- Adjusted JSON sidecar file naming convention to use `_meta.json` suffix.
- Updated relevant CMakeLists to include new source files and dependencies.
- Enhanced error handling and logging across components.

+++ main/display_renderer.c
@@ -287,7 +287,7 @@ esp_err_t display_renderer_set_rotation(display_rotation_t rotation)
         ESP_LOGW(DISPLAY_TAG, "Failed to set µGFX rotation: %s", esp_err_to_name(err));
     }
     
-    config_store_set_rotation(rotation);
+    config_store_set_rotation((uint16_t)rotation);
     g_rotation_in_progress = false;
     
     ESP_LOGI(DISPLAY_TAG, "Screen rotation set to %d degrees", rotation);
